# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Types.ObjectId": {
				"type": "string"
			},
			"StatusType": {
				"type": "string",
				"enum": [
					"Running",
					"Alerting",
					"Stopped"
				]
			},
			"AssetDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"serialCode": {
						"type": "string"
					},
					"manufacturer": {
						"type": "string"
					},
					"purchaseDate": {
						"type": "string",
						"format": "date-time"
					},
					"nextMaintenanceDate": {
						"type": "string",
						"format": "date-time"
					},
					"needMaintenance": {
						"type": "boolean"
					},
					"unitId": {
						"$ref": "#/components/schemas/Types.ObjectId"
					},
					"ownerId": {
						"$ref": "#/components/schemas/Types.ObjectId"
					},
					"model": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/StatusType"
					},
					"health": {
						"type": "number",
						"format": "double"
					},
					"imageUrl": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"serialCode",
					"manufacturer",
					"purchaseDate",
					"needMaintenance",
					"unitId",
					"ownerId",
					"model",
					"status",
					"health",
					"imageUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_AssetDto.Exclude_keyofAssetDto.id-or-unitId-or-ownerId-or-needMaintenance__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"serialCode": {
						"type": "string"
					},
					"manufacturer": {
						"type": "string"
					},
					"purchaseDate": {
						"type": "string",
						"format": "date-time"
					},
					"nextMaintenanceDate": {
						"type": "string",
						"format": "date-time"
					},
					"model": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/StatusType"
					},
					"health": {
						"type": "number",
						"format": "double"
					},
					"imageUrl": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"serialCode",
					"manufacturer",
					"purchaseDate",
					"model",
					"status",
					"health",
					"imageUrl"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_AssetDto.id-or-unitId-or-ownerId-or-needMaintenance_": {
				"$ref": "#/components/schemas/Pick_AssetDto.Exclude_keyofAssetDto.id-or-unitId-or-ownerId-or-needMaintenance__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateAssetDto": {
				"$ref": "#/components/schemas/Omit_AssetDto.id-or-unitId-or-ownerId-or-needMaintenance_"
			},
			"Partial_CreateAssetDto_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"serialCode": {
						"type": "string"
					},
					"manufacturer": {
						"type": "string"
					},
					"purchaseDate": {
						"type": "string",
						"format": "date-time"
					},
					"nextMaintenanceDate": {
						"type": "string",
						"format": "date-time"
					},
					"model": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/StatusType"
					},
					"health": {
						"type": "number",
						"format": "double"
					},
					"imageUrl": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateAssetDto": {
				"$ref": "#/components/schemas/Partial_CreateAssetDto_"
			},
			"CompanyDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"industry": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_CompanyDto.Exclude_keyofCompanyDto.users-or-units-or-_id-or-id__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"industry": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_CompanyDto.users-or-units-or-_id-or-id_": {
				"$ref": "#/components/schemas/Pick_CompanyDto.Exclude_keyofCompanyDto.users-or-units-or-_id-or-id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateCompanyDto": {
				"$ref": "#/components/schemas/Omit_CompanyDto.users-or-units-or-_id-or-id_"
			},
			"Partial_CreateCompanyDto_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"industry": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateCompanyDto": {
				"$ref": "#/components/schemas/Partial_CreateCompanyDto_"
			},
			"RoleType": {
				"type": "string",
				"enum": [
					"admin",
					"technician"
				]
			},
			"UserDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"companyId": {
						"$ref": "#/components/schemas/Types.ObjectId"
					},
					"role": {
						"$ref": "#/components/schemas/RoleType"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"companyId",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserDto": {
				"properties": {
					"role": {
						"$ref": "#/components/schemas/RoleType"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name"
				],
				"type": "object"
			},
			"Partial_CreateUserDto_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/RoleType"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateUserDto": {
				"$ref": "#/components/schemas/Partial_CreateUserDto_"
			},
			"UnitDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"companyId": {
						"$ref": "#/components/schemas/Types.ObjectId"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"location": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"companyId",
					"name",
					"location"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUnitDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"location": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"location"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_CreateUnitDto_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"location": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateUnitDto": {
				"$ref": "#/components/schemas/Partial_CreateUnitDto_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "tractian-challenge",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/assets?low_life": {
			"get": {
				"operationId": "lisLowHealthAssets",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AssetDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"assets"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "lowHealthThreshold",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/assets?needing_maintenance": {
			"get": {
				"operationId": "lisLowHealthAssets",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AssetDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"assets"
				],
				"security": [],
				"parameters": []
			}
		},
		"/units/{unitId}/assets": {
			"get": {
				"operationId": "listAssets",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AssetDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"assets"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "unitId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/StatusType"
						}
					}
				]
			},
			"post": {
				"operationId": "createEmployee",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AssetDto"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Pressure Sensor",
											"description": "Miniature Barometer",
											"health": 72,
											"imageUrl": "https://upload.wikimedia.org/wikipedia/commons/c/cb/Digital_Barometer_Sensor.jpg",
											"manufacturer": "Sensors SA",
											"model": "sensor x",
											"purchaseDate": "2023-04-15T01:13:40.879Z",
											"serialCode": "abc-123",
											"status": "Running"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"assets"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "unitId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAssetDto"
							}
						}
					}
				}
			}
		},
		"/units/{unitId}/assets/{assetId}": {
			"get": {
				"operationId": "getAsset",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AssetDto"
								}
							}
						}
					}
				},
				"tags": [
					"assets"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "unitId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					},
					{
						"in": "path",
						"name": "assetId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			},
			"patch": {
				"operationId": "updateAsset",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AssetDto"
								}
							}
						}
					}
				},
				"tags": [
					"assets"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "unitId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					},
					{
						"in": "path",
						"name": "assetId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAssetDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "deleteAsset",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"assets"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "unitId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					},
					{
						"in": "path",
						"name": "assetId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			}
		},
		"/companies": {
			"get": {
				"operationId": "listCompanies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CompanyDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"companies"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "industry",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "createCompany",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CompanyDto"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Industria Freios Supremos",
											"description": "Freios e ABS.",
											"industry": "auto parts manufacturer"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"companies"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCompanyDto"
							}
						}
					}
				}
			}
		},
		"/companies/{companyId}": {
			"patch": {
				"operationId": "createCompany",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CompanyDto"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Industria Freios Supremos",
											"description": "Freios e ABS.",
											"industry": "auto parts manufacturer"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"companies"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCompanyDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "deleteCompany",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"companies"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			},
			"get": {
				"operationId": "getCompany",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CompanyDto"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"404": {
						"description": "Company Not Found"
					}
				},
				"tags": [
					"companies"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			}
		},
		"/companies/{companyId}/employees": {
			"get": {
				"operationId": "listEmployees",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"employees"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			},
			"post": {
				"operationId": "createEmployee",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "John Doe",
											"email": "john.doe@mail.com",
											"role": "technician"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"employees"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			}
		},
		"/companies/{companyId}/employees/{userId}": {
			"get": {
				"operationId": "getEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"employees"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			},
			"patch": {
				"operationId": "updateEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"employees"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "deleteEmployee",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"employees"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			}
		},
		"/companies/{companyId}/units": {
			"get": {
				"operationId": "listUnits",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UnitDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"units"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			},
			"post": {
				"operationId": "createEmployee",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnitDto"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Unit 02",
											"location": "New Tokyo",
											"description": "Advanced Technology"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"units"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUnitDto"
							}
						}
					}
				}
			}
		},
		"/companies/{companyId}/units/{unitId}": {
			"get": {
				"operationId": "getUnit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnitDto"
								}
							}
						}
					}
				},
				"tags": [
					"units"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					},
					{
						"in": "path",
						"name": "unitId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			},
			"patch": {
				"operationId": "updateUnit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnitDto"
								}
							}
						}
					}
				},
				"tags": [
					"units"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					},
					{
						"in": "path",
						"name": "unitId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUnitDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "deleteUnit",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"units"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "companyId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					},
					{
						"in": "path",
						"name": "unitId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Types.ObjectId"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}